lint: remove end-of-line whitespace
This patch removes extraneous end-of-line whitespace, while preserving
the double space at the end of lines that indicates a paragraph break.

Additionally, a linting rule is added to ensure that these spaces stay
away. This should help prevent the cases where a single space is added,
where a double space is meant, as well as the case where accidental
spaces are left at the end of lines.

Signed-off-by: Justin Chadwell <me@jedevc.com>